aplikacja/requirements.txt
cffi==1.17.1
cryptography==44.0.0
numpy==2.2.0
opencv-python==4.10.0.84
pycparser==2.22
pycryptodome==3.21.0
pyOpenSSL==24.3.0
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
shiboken6==6.8.1
